Fixing main menu infinite loop and replay button

+++ Assets/Takeoff.cs
@@ -20,7 +20,7 @@ public class Takeoff : MonoBehaviour
 
         int startVelocity = 1050;
 
-        int playersLeft = (int) (totalScore / 50);
+        int playersLeft = totalScore / 50;
         GameState.numSurvivors = playersLeft;
 
         foreach (GameObject trainCar in GameObject.FindGameObjectsWithTag("Train"))
